requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
biopython==1.73
boto==2.49.0
boto3==1.9.234
botocore==1.12.234
bz2file==0.98
certifi==2020.4.5.1
cffi==1.12.3
chardet==3.0.4
click==7.1.2
conda==4.7.12
conda-package-handling==1.3.11
cryptography==2.7
docutils==0.15.2
Flask==1.1.2
gensim==3.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.4
lazy-object-proxy==1.4.1
libarchive-c==2.8
MarkupSafe==1.1.1
menuinst==1.4.16
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
numpy==1.16.4
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.0
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.3
pywin32==223
requests==2.22.0
ruamel-yaml==0.15.46
s3transfer==0.2.1
scipy==1.3.1
six==1.12.0
smart-open==1.8.4
tqdm==4.32.1
typed-ast==1.3.5
urllib3==1.24.2
virtualenv==16.7.5
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.1
